Zmiana weryfikowania uprawnień

+++ blog/src/app/components/blog-components/blog-item/blog-item.component.ts
@@ -1,19 +1,19 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { CommonModule } from "@angular/common";
 import { BlogItemImageComponent } from "../blog-item-image/blog-item-image.component";
 import { BlogItemTextComponent } from "../blog-item-text/blog-item-text.component";
-import { AuthService } from '../../../services/auth/auth.service';
 import { DataService } from "../../../services/post/post.service";
+import { AuthService } from '../../../services/auth/auth.service';
 
 @Component({
   selector: 'blog-item',
   standalone: true,
   imports: [BlogItemImageComponent, BlogItemTextComponent, CommonModule],
-  providers: [AuthService],
+  providers: [DataService],
   templateUrl: './blog-item.component.html',
-  styleUrl: './blog-item.component.css'
+  styleUrls: ['./blog-item.component.css']
 })
-export class BlogItemComponent {
+export class BlogItemComponent implements OnInit {
   @Input() title?: string;
   @Input() image?: string;
   @Input() text?: string;
@@ -22,20 +22,20 @@ export class BlogItemComponent {
   public isAdmin?: boolean;
 
   constructor(
-    private authService: AuthService,
-    private dataService: DataService) { }
+    private dataService: DataService,
+    private authService: AuthService) { }
 
   ngOnInit(): void {
-    if (this.authService.isLoggedIn()) {
-      this.isAdmin = this.authService.isAdmin();
-    }
+    this.isAdmin = this.authService.isAdmin();
   }
 
   removePost() {
-    if (this.id != null) {
+    if (this.id != null && this.isAdmin) {
       this.dataService.removePost(this.id).subscribe(response => {
-      }, error => {});
-    }
-    window.location.reload(); 
+      }, error => {
+        console.error("Error removing post:", error);
+      });
+      window.location.reload();
+    } 
   }
 }
